requirements.txt
openai>=1.12.0
supabase>=2.0.0
tiktoken>=0.5.0
pydantic>=2.0.0
tenacity>=8.2.0
python-dotenv>=1.0.0
streamlit>=1.37.0
streamlit-authenticator==0.4.2
plotly>=5.18.0
psycopg2-binary>=2.9.0
mcp>=1.2.0
